+++ src/circularProgress/index.tsx
@@ -1,26 +1,13 @@
-import React, { useEffect, useMemo } from 'react';
-import {
-  StyleSheet,
-  Text,
-  TextInput,
-  View,
-} from 'react-native';
-import Svg, { G, Circle, Defs, LinearGradient, Stop } from 'react-native-svg';
-import Animated, {
-  useSharedValue,
-  withTiming,
-  useAnimatedProps,
-  withDelay,
-  runOnJS,
-  useDerivedValue,
-  Easing,
-} from 'react-native-reanimated';
+import React, { useMemo } from 'react';
+import { Text, TextInput, StyleSheet, View } from 'react-native';
+import Animated from 'react-native-reanimated';
+import ProgressCircle from '../components/progressCircle';
+import useAnimatedValue from '../hooks/useAnimatedValue';
 import COLORS from '../utils/colors';
 import styles from './styles';
-import {CircularProgressProps} from './types';
+import { CircularProgressProps } from './types';
 
 const AnimatedInput = Animated.createAnimatedComponent(TextInput);
-const AnimatedCircle = Animated.createAnimatedComponent(Circle);
 
 const CircularProgress: React.FC<CircularProgressProps> = ({
   value,
@@ -60,9 +47,19 @@ const CircularProgress: React.FC<CircularProgressProps> = ({
     return Math.round(v);
   },
 }: CircularProgressProps) => {
-  const animatedValue = useSharedValue(initialValue);
-  const viewBox = radius + Math.max(activeStrokeWidth, inActiveStrokeWidth);
-  const circleCircumference = 2 * Math.PI * radius;
+  const { animatedCircleProps, animatedTextProps } = useAnimatedValue({
+    initialValue,
+    radius,
+    maxValue,
+    clockwise,
+    delay,
+    value,
+    duration,
+    onAnimationComplete,
+    valuePrefix,
+    progressFormatter,
+    valueSuffix,
+  });
 
   const styleProps = useMemo(
     () => ({
@@ -97,82 +94,21 @@ const CircularProgress: React.FC<CircularProgressProps> = ({
     ]
   );
 
-  const animatedCircleProps = useAnimatedProps(() => {
-    let biggestValue = Math.max(initialValue, maxValue);
-    biggestValue = biggestValue <= 0 ? 1 : biggestValue;
-    const maxPercentage: number = clockwise
-      ? (100 * animatedValue.value) / biggestValue
-      : (100 * -animatedValue.value) / biggestValue;
-    return {
-      strokeDashoffset:
-        circleCircumference - (circleCircumference * maxPercentage) / 100,
-    };
-  });
-
-  useEffect(() => {
-    animatedValue.value = withDelay(
-      delay,
-      withTiming(value, { duration, easing: Easing.linear }, (isFinished) => {
-        if (isFinished) {
-          runOnJS(onAnimationComplete)?.();
-        }
-      })
-    );
-  }, [value]);
-
-  const progressValue = useDerivedValue(() => {
-    return `${valuePrefix}${progressFormatter(
-      animatedValue.value
-    )}${valueSuffix}`;
-  });
-
-  const animatedTextProps = useAnimatedProps(() => {
-    return {
-      text: progressValue.value,
-    } as any;
-  });
-
   return (
     <View style={styles(styleProps).container}>
       <View style={styles(styleProps).rotatingContainer}>
-        <Svg
-          width={radius * 2}
-          height={radius * 2}
-          viewBox={`0 0 ${viewBox * 2} ${viewBox * 2}`}
-        >
-          {activeStrokeSecondaryColor ? (
-            <Defs>
-              <LinearGradient id="grad" x1="0%" y1="0%" x2="0%" y2="100%">
-                <Stop offset="0%" stopColor={activeStrokeSecondaryColor} />
-                <Stop offset="100%" stopColor={activeStrokeColor} />
-              </LinearGradient>
-            </Defs>
-          ) : null}
-          <G rotation="270" origin={`${viewBox}, ${viewBox}`}>
-            <Circle
-              cx="50%"
-              cy="50%"
-              stroke={inActiveStrokeColor}
-              strokeWidth={inActiveStrokeWidth}
-              r={radius}
-              fill={circleBackgroundColor}
-              strokeOpacity={inActiveStrokeOpacity}
-            />
-            <AnimatedCircle
-              cx="50%"
-              cy="50%"
-              stroke={
-                activeStrokeSecondaryColor ? 'url(#grad)' : activeStrokeColor
-              }
-              strokeWidth={activeStrokeWidth}
-              r={radius}
-              fill="transparent"
-              strokeDasharray={circleCircumference}
-              animatedProps={animatedCircleProps}
-              strokeLinecap={strokeLinecap}
-            />
-          </G>
-        </Svg>
+        <ProgressCircle
+          circleBackgroundColor={circleBackgroundColor}
+          radius={radius}
+          strokeLinecap={strokeLinecap}
+          activeStrokeColor={activeStrokeColor}
+          activeStrokeSecondaryColor={activeStrokeSecondaryColor}
+          activeStrokeWidth={activeStrokeWidth}
+          inActiveStrokeColor={inActiveStrokeColor}
+          inActiveStrokeWidth={inActiveStrokeWidth}
+          inActiveStrokeOpacity={inActiveStrokeOpacity}
+          animatedCircleProps={animatedCircleProps}
+        />
       </View>
       <View
         style={[